+++ guillotina_stripe/tests/fixtures.py
@@ -4,18 +4,20 @@ from guillotina import testing
 import pytest
 import os
 
+
 def base_settings_configurator(settings):
     if 'applications' in settings:
         settings['applications'].append('guillotina_stripe')
         settings['applications'].append('guillotina_stripe.test_package')
     else:
-        settings['applications'] = ['guillotina_stripe', 'guillotina_stripe.test_package']
+        settings['applications'] = [
+            'guillotina_stripe', 'guillotina_stripe.test_package']
     settings['load_utilities']['stripe'] = {
         "provides": "guillotina_stripe.interfaces.IStripePayUtility",
         "factory": "guillotina_stripe.utility.StripePayUtility",
         "settings": {
-            "secret": os.environ.get("STRIPE_KEY", ""),
-            "signing": os.environ.get("SIGNING", ""),
+            "secret": os.environ.get("STRIPE_KEY", "sk_test_51ALBplGeGvgK89lRFefDLO25NO72Q8NbSR4MHAaAQpB9MuoVGWW61va9V8dLt6IqQgDSKrt2tMgIxozrMtmLbop700dcPnjHda"),
+            "signing": os.environ.get("SIGNING", "pk_test_XSVr7Cj8yI8kG7HJBVI0kqqx"),
             "testing": True
         },
     }
@@ -36,6 +38,7 @@ settings['stripe'] = {
     'image': 'stripemock/stripe-mock',
     'version': 'latest',
 }
+
 
 class STRIPE(BaseImage):
     name = 'stripe'
@@ -59,11 +62,9 @@ class STRIPE(BaseImage):
 
 stripe_image = STRIPE()
 
+
 @pytest.fixture(scope='session')
 def stripe():
     result = stripe_image.run()
     yield result
     stripe_image.stop()
-
-
-