+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ CHANGELOG
 ---------------------
 
 - Not delete consumer after cancel subscription [rboixaderg]
+- Add cards behavior. Create a new behavior to save all customer cards in object. [rboixaderg]
+- Not cancelled all subscriptions from the same customer when created a new subscription or delete it [rboixaderg]
 
 
 1.0.0a11 (2021-11-25)